requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
psycopg2-binary==2.9.9
requests==2.31.0
celery==5.3.6
redis==5.0.1
drf-yasg==1.21.7
pytest-django==4.7.0
polyline==1.4.0